requirements.txt
numpy
tensorflow
scikit-learn
pandas
matplotlib
Flask
flask-cors
Flask-SQLAlchemy
opencv-python
requests
Flask-Login
Flask-SocketIO
python-socketio
Werkzeug
python-engineio
eventlet
psycopg2-binary
python-dotenv